add whitelist capability
A whitelist is helpful if you want to run the tourney scripts
for a small group.

If WHITELISTFILE is set to a valid/existing file, then is_not_tourney()
returns True for players not in the whitelist file.

WHITELISTFILE defaults to 'player-whitelist.txt'. Simply omit this
file if you don't want to use the whitelist.