Fix handling of non-ascii characters in Mako
Logfile/milestone entries with non-ascii were causing Mako to
raise an exception. The workaround in this commit imports a
filter function that converts all string data to unicode and
applies this filter to all values.  The Mako docs say that
unicode is the default filter applied to data, but that may not
be the case for the version we have installed, or in any case
isn't sufficient. The filter is based on the suggestion here:

http://stackoverflow.com/a/25235213