Set the clan deadline for 11PM UTC so it can update clans
Due to an error on my point, the last RC file fetch was schedules 2h before the
deadline, so some RCs were not fetched. Hence we're extending the clan cutoff
to 10pm UTC so that RCs will fetch and then setting the deadline in the script
to 11pm UTC. Also update the tournament guide to mention this issue, so we
hopefully won't make that mistake again.