Implement a player blacklist
Add support to the scripts for a player blacklist. If the file
player_blacklist.txt exists, its lines are loaded as player names into a
player_blacklist list. Any logfiles or milestones by that player or
related to that player's ghost are ignored.

Additionally there's a blacklist-player-db.sh script that can remove any
current records associated with a player from the database. This is
necessary to prevent the scripts from producing output about already
processed data. The tournament guide now has instructions on how to use
the blacklist.

This commit also renames the blacklist.txt file to game_blacklist.txt to
make it clear that this blacklist is for removing specific games only.

TODO: I doubt that kills of the blacklisted player's ghost are properly
ignored, but I want to get the blacklist going in the current tourney.