Update unicode handling for templates
I'm not sure really what the old logic was for or why it was created. It
seemed to assume templates were ascii. But if you added unicode to a
template and marked it as unicode (with a -*- encoding line up the top),
it wouldn't work...

Throw out all the old logic and support Python 2 style unicode. That is,
all strings are ascii unless prefixed with 'u'. Which is annoying to
remember, but works for now.