Add new rules as Python data
Some of these are incomplete (individual challenges for first win and
first 15-rune win, banner for wu jian, banner notes for yred. This file
could also do with some load-time assertions on the data -- eg checking
that every god has a banner, and that there are no duplicate gods in the
individual challenge list (if that ends up being a valid constraint?!).