Remove old code from the points-based scoring system
This is largely about the pleasure of deleting lots of cruft. New
scoring does not need point-audit-trails, the system is based entirely
on player ranks that can be computed from the raw game and milestone
data and don't depend on time-sensitive scoring formulae.

Barring a complete re-write someone will have to drag this codebase
kicking and screaming to the python3 era at some point and this deletion
is my gift to them.