Compute ranks and overall scores, enable some old pages with tweaks
In principle the ranking and overall scores can be computed from the
database state with pure reads. In practice expressing the player_ranks
table as a view results in a very heavyweight view to be computed for
queries, because the use of window functions requires computing the
whole column just to get a single row. So, we store the ranks.

The overall score could at this point be a generated column in the DB
(!) but it's in python for now for prototyping.