Convert html.table_text to bootstrap
This will look ugly on old-style pages.

New style tables are much bulkier. It remains to be seen if this is a
problem, and if so how to address it.

I'd rather not fix it by making the table font smaller -- the
readability of the tables is already a little low.