ebering's fixed local test data for new scoring
This test configuration assumes you've already got the cpo log and
milestones downloaded, and runs over the interval of the first half of
the 0.24 tournament. To ensure best results, copy the archived nemelex
combos to nemelex-combos.txt

This data range has something nontrivial in every scoring category but
is small enough to run quickly on local rebuild.