Refactor configuration variable logic
Default LOCAL_TEST to be disabled, since we're using docket to test
things now. In the future, we might be able to remove this logic, but
I'm leaving it in for now.

Refactor the environment variable handling to not use elaborate boolean
logic and to generally defer to the relevant environment variables even
under LOCAL_TEST=true.