Some code fixes for CDO
The MySQLdb module on CDO requires a slightly different connection
argument for the password. Additionally the exceptions are found
directly in the module instead of through an '_exceptions' member. The
password argument seems particular to the somewhat older version of the
module CDO uses, but the exception locations seem consistent with the
current documentation of this module. After tournament we'll do an
upgrade to python3 and will have a more current module, so it will be
more clear which of these changes we'll either keep or revert.