Fix clan page generation on CDO (gammafunk)
The previous formulation works in Docker, but not the real production
environment in CDO. I believe this is because the MySQL module used on
CDO is different and isn't smart enough to handle list-of-string
insertion into the query (it seems to be running repr() on the list and
inserting that, which is totally wrong).

Instead of relying on list quoting at all, manually expand the list into
individual values.