Add a CLO to not fetch/udpate db
Convenience option for testing page updates. Same name as the comparable
option in `scoring`. It would also be nice to be able to not do the full
player page update cycle, but annoyingly, this is all triggered
indirectly by listeners so is a bit complex to patch in compared to how
this is done in scoring.