Add player/clan search
This is a long-requested feature, but maybe hasn't been entirely obvious
how to implement given that we are doing static site generation. It is
implemented entirely in javascript: search.html is generated with a big
table that has every player and clan name, and the search boxes filter
these down. There are no doubt many further tweaks that could be added
here, and due to ignorance I did not use bootstrap at all for styling
this, but this gets the basic functionality. The size of search.html is
about 128K, which is reasonable in the scale of file sizes we already
have.