Use DataTables for all-players, teams pages
This enables sorting by arbitrary columns, as well as searching in these
tables. There is still a bunch more tweaking to be done. DataTables
supports pagination but I didn't find it easy to use with these tables,
so I've held off on enabling it.

TODO: clans sort players with a clan first.