Update clan page URLs for the final time
The new format is ${team_name}-${captain}.html. Which is the same as
before, but now the whole url is lowercased and slugified.

As part of this, fix crawl_utils.clan_link and use it everywhere. This
removes a whole bunch of reimplementations spread throughout the
codebase (mostly added by me 😊).