Update the set of eligible Nemelex Choice combos
Unfortunately, with Sequell down I can't run the necessary query to
determine what set of combos we want. I'm simply modifying the existing
file to remove nem choice combos from this file that were won in the
last tournament. This still gives us 82 combos to choose from, which is
a decent amount of variety. Hopefully we can get the necessary info to
update this file after tournament starts.