Tweak some things and use more Magenta
Green for wins worked well in CSDC because there were three states to
distinguish in a single table: ongoing, won, and lost, and the CSDC
sites style was overall purple for that Zotfeel.

The tournament site uses a neutral grey theme, and ongoing games are
always displayed separate from completed games, so we can use magenta
for win highlighting for that sweet sweet Zotfeel.

[ Closes #11 ]