Sort "" and "-" last
This is done by using the 'data-sort' attribute at generation time.
Several elements of this are kind of rough, especially how columns with
links in them are handled.

If someone makes a player or team named with a bunch of Zs, this will
get messed up, but I'm not sure of any other way to handle it on this
approach. (An alternative option would be to write custom sort
functions in js, but this is even more involved given how our tables are
generated.)