Add a wins/top games leaderboard for clans
This may add too much vertical space for people's preferences, and for
teams with few games there's some redundancy. However, teams really do
need to be able to see all their wins in one place, and for teams with
few wins I think it is nice to see a score-sorted leaderboard.