Switch to DENSE_RANK()
For scoring purposes, using 10,000 / rank for points in ranked
categories works great, but the resolution of ties for ranking leaves
something to be desired. In practice in each of these categories we're
interested in rating the relative player skill against the field,
independent of field size, so DENSE_RANK is more appropriate.

For example, a low XL win at XL 25 should not be worth less depending on
the number of XL 24 wins, but should be if someone adds an XL 23 win.