Improve some handling of timeout issues
If `wget` failed entirely, this code threw an exception that wasn't
caught except at the global level. This effectively aborts the reload
loop for any servers after the failing one. I've just changed this to
log an error, which should be fine.

Also, drop the timeout / retry count somewhat. The default number of
retries is 20, so for a server that is actually timing out, the old
values would amount to a 20 minute wait.