Reset the scripts for the next tournament
Re-enable test mode and comment out all sources in both test and live
data, since we enabled the sources incrementally as the branches are
installed.

Ideally this data would be moved to configuration files that lived
outside the repo. We could then store the templates in the repo,
updating those as new options were added. As it stands, we have to
commit every configuration change.