feat: Update the list of Nemelex Choice combos
The list of eligible Nemelex Choice combos are those given by the
Sequell command:

!lg * won playable !experimental !nchoice s=char o=-char format:"${.}" join:"|" title:"" ?:N<=52

This gives a set of 117 combos. Also update the Nemelex Choice banner
description to indicate how the random combos are chosen. Not sure if
the updated text will fit, so I may need to find a way to trim it if
it's too long.