feat: Improve Nemelex Choice combo listing
When the combo listing gets long, it's very hard to read the actual
combos in the long block of text, even with the existing spacing and
punctuation. Make the combos themselves bold, which should hopefully
improve things. It might be better to try other types of formatting or
even a different way of displaying this data, but this will do for now.