feat: Re-retweak the Nemelex Choice combo display
The <code> tag is in red, but uses a smaller font. Apparently the
correct way to get bold is with a span tag, as evidenced by the
tournament date display, so let's try this. Likewise update the bolding
in the listing of previous combos, since that would have the same issue.