feat: remove the first win categories
These don't score well as relative categories. They also are a source of
vocal frustration from the players who can't start on time for the
tournament. The first win tables are kept in-tact because it would be
nice to shout-out the win order on the tournament page (even if not for
score), but I don't think I'll have time to finish that before 0.28 T.