fix: score rounding differences between pages
The overview page was sometimes showing a slightly different player
total score compared to the value shown on the player page. This was
because we recalculate it on the latter in the template, whereas the
former uses the value directly from the DB, and the template is
performing slightly different rounding.

Instead of recalculating the overall player score, pass it as a value to
the player page template along with the rest of the data we already
pass.