feat: Restrict Meteorans from class high scores
Meteorans are a very dominant species in terms of high score, more so
than any species that has existed before. This means that setting a
class high score with a non-Meteoran during a long tournament would
likely be very difficult. To help ensure variety for the class high
scores, we're preventing Meteorans from setting them. This means that
meteoran games can't claim the 27 point bonus for setting a class high
score, nor can a player get the Dithmenos 3 banner specifically when
stealing a class high score.

Meteorans are still eligible for the overall low turncount and high
score rankings so that the speedrun-inspired species retains good
representation among speedrun challenges. Meteorans can also set their
combo highscores and the meteoran species high score normally. We've
talked about finally moving to aut-based scoring for 0.30, so it's
possible that this restriction may not be necessary for the 0.30
tournament.