task: Add CDI's bot-only sources to the tournament
We're running qw on CDI on an install that's only accessible to
specially designated bot users. Hence we'll be able to allow qw to
participate in the tournament without causing any burden to official
servers. Add logfile and milestones urls for this installation. A
subsequent commit will ensure that qw doesn't recieve a ranking for the
realtime category.