feat: Exclude bots from the realtime rankings
Add clauses to exclude the list of known bots to the views that return
the realtime rankings. This approach isn't dynamic in that account names
need to be coded directly into the view SQL, but this makes it easy
implement without having to change various areas of the python code, and
we don't expect there to be a large number of registered bots.