feat: Exclude bots from realtime and stats
Add clauses to exclude known bots from the views that return the
realtime rankings. This approach isn't dynamic in that account names
need to be coded directly into the view SQL, but this makes it easy
implement without having to change various areas of the python code, and
we don't expect there to be a large number of registered bots.

We also exclude bots from the overall tournament stats. Bots are allowed
to otherwise earn points in other categories and have an overall player
ranking. If we'd like to exclude bots from other categories, it's
straightforward to do so either via the relevant view or by using the
newly added variable defining such bots. Currently we mark qw and
tstbtto as known bots. The latter hasn't played since 2021, but that's
recent enough to warrant pre-emptive exclusion.