fix: fix two painfully slow views
On tournament-scale data, both of the replaced queries take a very long
time on mysql on both cdo and in my local tests (currently 3-5s locally
for the player score, and several minutes for the clan score). This
results in massively slow player/clan page generation. It's actually
somewhat unclear to me why, but in any case, the necessary views can be
constructed also using an arguably easier-to-understand inner join with
a 'GROUP BY'+MAX -style query. In my local tests (on a faster machine
than cdo for sure) a full reload cycle with current data takes about 5-6
minutes, and a taildb loop is extremely fast.