fix: calculate heretic 1 correctly
This query had two bugs. First, game_id appears to have been put in the
milestones table as a column but never implemented. Second, the query
did not correctly handle games where abandonment was the last milestone
(game ending events don't generate a milestone): the left outer join in
that case is guaranteed to be null. This commit is more elaborate but
more direct about what it is calculating, and should fix up those cases.