fix: tweak Makhleb 1
Makhleb's tier 1 banner was essentially impossible for anyone who
couldn't do tier 2 also. This commit doubles the required time, which
will make it substantially easier. Perhaps too easy; the median time
to D:15 in the last tournament was around 75 minutes, so this is still
well below that, but this change may well put it within reach of most
players who can get to D:15 at all.