fix: tweak vehumet banner
This commit adjusts the vehumet / ruthless efficiency tiers to make it
somewhat doable, but probably still one of the harder banners. This was
nearly the least achievable banner overall in t0.29, with just a single
person getting the tier 3. Even with the changes, only 22 games in t0.29
would be eligible for the new tier 3.