task: The eligible Nemelex Choice combos for 0.30
As given by the following query:

!lg * won playable !experimental !nchoice-0.29 s=char o=-char format:"${.}" join:"|" title:"" ?: N<=60

The max number of wins for a combo could actually decrease this year
from 54 to 51 if we wanted at least 110 combos. This is because we've
replaced the Pa species with At, which has fewer wins, and replaced the
AK class with Re, which has far fewer wins. However this would leave us
with only 25 combos out of 111 (22%) that don't involve the two least
won species (At and Me) or two least won classes (Re and CA). The
selection algorithm down-weightings combos by how often their
species/class has already been selected, but it can't do anything about
a lack of variety in the combo pool itself.

To account for this, I've increased the number of wins required from 57
to 60, which gives us 143 combos. Of these, 56 (39%) don't involve
At/Me/Re/CA, so we'll have better combo variety.