fix: Correct info on Nemelex Choice combos
I *think* we used a cutoff of 49 wins for the eligible combos. One of
the two descriptions had a cutoff of 52 wins, but the previous commit
mentions 49 in the command and the stated number of resulting combos
matches the number in nem_eligible.txt, so we'll go with that. I'll
manually update the generated html for 0.28.