fix: Scoring for Ziggurats past 27 (RoGGa, Ge0ff)
We were capping the value of total Ziggurat floors completed at
28 * 78 = 756 in the category page. However this isn't enough to prevent
values greater than the cap from being scored, since the relevant
calculations just pull values from the view. So this commit caps the
value returned for `floors` to 756 without touching the number of zigs
completed and deepest floors. This way we can display info on zigs
completed past 27 but not score past the cap.