feat: A Gem category and banner
Since 0.31 has gems, which reward the player for reaching the end of
branches quickly in terms of turncount. The gem collection category is proportional
and gives players points based on how many of the 13 types of gems
they've collected in any of their games. The clan gem collection
category is simply the same ranking but considering the clan's games as
a group. Uskayaw also has banners for finding a gem, winning with 3 gems
intact, and winning with all 11 gems intact.