task: The eligible Nemelex Choice combos for 0.31
As given by the following query:

!lg * won playable !experimental !nchoice-0.30 s=char o=-char format:"${.}" join:"|" title:"" ?: N<=64

This increases the number of required wins from 60 to 64. This gives 69
out of 151 combos not involving the two species (On and At) and two
classes (Al and Sh) with the most combos. Hence there should be good
combo variety overall, and of course the selection algorithm will make
sure no species and class is selected too often.