feat: A game allow list (Napkin)
Currently the game blocklist allows matching games via arbitrary xlog
lines at the time they're processed. This allows match flexibility, but
generally speaking we only need to block games by name and server. The
block list didn't set the src field that gives the server, so previously
when dealing with e.g. griefers, we had to match some other unique data
in the game (in addition to username). Furthermore griefers may attempt
to register a user on multiple servers, so the ideal implementation is a
conditional allow list: if a user has no entries, all their games are
allowed, and if a user does have an entry, only games on the indicated
servers would be allowed.

This commit implements this form of allow list, additionally setting the
src field before checking our list, so we can do server matches. The new
format is:

NAME:SRC1,SRC2,...

where NAME is the player name, and the SRC entries are server
abbreviations where games for this player are allowed. Matching is case
insensitive, and it works for both logfiles and milestone files. When
games are blocked, instead of "munging" records to have fake data, we
skip them, since there's no need to keep these records. The allowlist
file is game_allowlist.txt (renamed from game_blocklist.txt).

I've tested it locally with appropriate entries for the known griefed
users, and it works for removing the bad records with no apparent
problems. Hence I'll push this and update CDO with the appropriate
allowlist file.