feat: Lower the Zig category cap from 27 to 5
The Ziggurat Diving category is widely regarded as the most onerous and
boring tournament category. This is both because of how many zigs must
be cleared to reach maximum points and the fact that zig completion
doesn't feel particularly interesting past the first few. This commit
caps the scoring at 5 zigs, hopefully making this category much less
annoying to competitive players and clans. Zigs past 5 are still
displayed on the category summary pages, but only 5 will actually count
for player and clan scores.