fix: Handle Yred's `bound` milestone
Binding a monster's soul produces an additional uniq verb milestone in
addition the the standard one. Ignore the bound milestone by
transforming it into "uniq.bound" in the usual place for this,
xlog_milestone_fixup(), before it gets passed to any milestone handler.
Also don't shorten the names of similarly transformed milestone verbs,
since we have plenty of characters to work with and don't aggressively
shorten other milestone verbs. The VARCHAR size of the verb name can
always be increased, if need be.