Adjust Trog banner logic and descriptions
For Trog 3, in addition to the newly added requirements that other rune
branches not be entered, also require that the player hasn't visited
Vaults:5 that game. This prevents them from clearing Vaults:5 and
leaving the rune on the ground to satisfy the previous conditions. It
doesn't make the banner meaningfully harder, but it prevents any
remaining encouragment of rune pickup avoidance.

Also optimize the relevant query by avoiding doing it as much as
possible and tweak the description language. Closes #16.