Yet more tweaks to Trog banner code and rules
For Trog 2, don't allow entering Tomb. The player wouldn't generally
want to do this, but for consistency with not allowing entering other
extended branches, it makes sense to disallow this. Optimize and
refactor the code a bit more, update the banner description, and mention
the Trog banner changes on the index page.