Remove stale combo highscore points
When a new combo high score is made, we have to invalidate the
combo_score value in the players table for the player who help the
previous high score. This is because this player may no longer have a
combo high score at all, in which case the UPDATE in
query.update_all_player_ranks() won't see any row to update. We're
already doing a DELETE to combo_highscores in this situation, so we just
add an UPDATE for the players table that sets their combo_score to 0.
For any players that do have at least one combo highscore, their valid
current score will get recalculated by query.update_all_player_ranks()
later on.